moved log_likelihood to infer module; function now returns dictionary of per-site likelihoods
